Create preliminary delete order sequence
Add button with id of order id to orders template.
Create event handler for deletion of order.
Make AJAX call to api to DELETE order.
Make UI notification when order is deleted

Need to define url for deletion as well as the data being sent
to the api (order id, user id, token)?